[ ADAM ] validation of adam optimizer
validation of adam

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/src/util_func.cpp
@@ -39,9 +39,9 @@ unsigned int getSeed() { return 0; }
 
 float random() { return dist(rng); }
 
-float sqrtFloat(float x) { return (float)(sqrt(x)); };
+float sqrtFloat(float x) { return sqrt(x); };
 
-float logFloat(float x) { return (float)(log(x)); }
+float logFloat(float x) { return log(x); }
 
 // This is 2D zero pad
 // TODO : Optimize for multi dimention padding
@@ -62,15 +62,15 @@ Tensor zero_pad(int batch, Tensor const &in, unsigned int const *padding) {
   for (unsigned int j = 0; j < c; ++j) {
     for (unsigned int k = 0; k < padding[0]; ++k) {
       for (unsigned int l = 0; l < width_p; ++l) {
-        output.setValue(0, j, k, l, 0.0);
-        output.setValue(0, j, k + height_p_h, l, 0.0);
+        output.setValue(0, j, k, l, 0.0f);
+        output.setValue(0, j, k + height_p_h, l, 0.0f);
       }
     }
 
     for (unsigned int l = 0; l < padding[1]; ++l) {
       for (unsigned int k = padding[0]; k < h; ++k) {
-        output.setValue(0, j, k, l, 0.0);
-        output.setValue(0, j, k, l + width_p_h, 0.0);
+        output.setValue(0, j, k, l, 0.0f);
+        output.setValue(0, j, k, l + width_p_h, 0.0f);
       }
     }
   }
